[1.0.0] [+]新增若干内容

+++ src/main/kotlin/com/github/cpjinan/plugin/akariartifact/internal/command/subcommand/LoreCommand.kt
@@ -8,21 +8,38 @@ import taboolib.common.platform.command.int
 import taboolib.common.platform.command.subCommand
 import taboolib.module.chat.colored
 import taboolib.module.lang.sendLang
+import taboolib.platform.util.isNotAir
 import taboolib.platform.util.modifyLore
 
 object LoreCommand {
     val clipboard: HashMap<Player, String> = hashMapOf()
 
     val lore = subCommand {
+        literal("check") {
+            execute<ProxyCommandSender> { sender: ProxyCommandSender, _: CommandContext<ProxyCommandSender>, _: String ->
+                sender.castSafely<Player>().let {
+                    val item = it?.inventory?.itemInMainHand
+                    if (item.isNotAir()) {
+                        sender.sendLang("Check-Lore")
+                        item.itemMeta?.lore?.forEachIndexed { index, content ->
+                            sender.sendMessage("&7${index + 1} &8| $content".colored())
+                        }
+                    } else sender.sendLang("Air-In-Hand")
+                }
+            }
+        }
+
         literal("add") {
             dynamic("lore") {
                 execute<ProxyCommandSender> { sender: ProxyCommandSender, context: CommandContext<ProxyCommandSender>, _: String ->
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            add(context["lore"].colored())
-                        }
-                        sender.sendLang("Add-Lore", context["lore"].colored())
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                add(context["lore"].colored())
+                            }
+                            sender.sendLang("Add-Lore", context["lore"].colored())
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }.dynamic("options") {
@@ -38,10 +55,12 @@ object LoreCommand {
 
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            add(context["lore"].colored())
-                        }
-                        if (!silent) sender.sendLang("Add-Lore", context["lore"].colored())
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                add(context["lore"].colored())
+                            }
+                            if (!silent) sender.sendLang("Add-Lore", context["lore"].colored())
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }
@@ -52,10 +71,12 @@ object LoreCommand {
                 execute<ProxyCommandSender> { sender: ProxyCommandSender, context: CommandContext<ProxyCommandSender>, _: String ->
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            removeAt(context.int("line") - 1)
-                        }
-                        sender.sendLang("Remove-Lore", context.int("line"))
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                removeAt(context.int("line") - 1)
+                            }
+                            sender.sendLang("Remove-Lore", context.int("line"))
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }.dynamic("options") {
@@ -71,10 +92,12 @@ object LoreCommand {
 
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            removeAt(context.int("line") - 1)
-                        }
-                        if (!silent) sender.sendLang("Remove-Lore", context.int("line"))
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                removeAt(context.int("line") - 1)
+                            }
+                            if (!silent) sender.sendLang("Remove-Lore", context.int("line"))
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }
@@ -85,10 +108,12 @@ object LoreCommand {
                 execute<ProxyCommandSender> { sender: ProxyCommandSender, context: CommandContext<ProxyCommandSender>, _: String ->
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            set(context.int("line") - 1, context["lore"].colored())
-                        }
-                        sender.sendLang("Set-Lore", context.int("line"), context["lore"].colored())
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                set(context.int("line") - 1, context["lore"].colored())
+                            }
+                            sender.sendLang("Set-Lore", context.int("line"), context["lore"].colored())
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }.dynamic("options") {
@@ -104,10 +129,12 @@ object LoreCommand {
 
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            set(context.int("line") - 1, context["lore"].colored())
-                        }
-                        if (!silent) sender.sendLang("Set-Lore", context.int("line"), context["lore"].colored())
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                set(context.int("line") - 1, context["lore"].colored())
+                            }
+                            if (!silent) sender.sendLang("Set-Lore", context.int("line"), context["lore"].colored())
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }
@@ -118,15 +145,17 @@ object LoreCommand {
                 execute<ProxyCommandSender> { sender: ProxyCommandSender, context: CommandContext<ProxyCommandSender>, _: String ->
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            add(context.int("line") - 1, context["lore"].colored())
-                        }
-                        sender.sendLang(
-                            "Insert-Lore",
-                            context.int("line"),
-                            context.int("line") + 1,
-                            context["lore"].colored()
-                        )
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                add(context.int("line"), context["lore"].colored())
+                            }
+                            sender.sendLang(
+                                "Insert-Lore",
+                                context.int("line"),
+                                context.int("line") + 1,
+                                context["lore"].colored()
+                            )
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }.dynamic("options") {
@@ -142,32 +171,36 @@ object LoreCommand {
 
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            add(context.int("line") - 1, context["lore"].colored())
-                        }
-                        if (!silent) sender.sendLang(
-                            "Insert-Lore",
-                            context.int("line"),
-                            context.int("line") + 1,
-                            context["lore"].colored()
-                        )
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                add(context.int("line"), context["lore"].colored())
+                            }
+                            if (!silent) sender.sendLang(
+                                "Insert-Lore",
+                                context.int("line"),
+                                context.int("line") + 1,
+                                context["lore"].colored()
+                            )
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }
         }
 
         literal("clear") {
-            execute<ProxyCommandSender> { sender: ProxyCommandSender, context: CommandContext<ProxyCommandSender>, _: String ->
+            execute<ProxyCommandSender> { sender: ProxyCommandSender, _: CommandContext<ProxyCommandSender>, _: String ->
                 sender.castSafely<Player>().let {
                     val item = it?.inventory?.itemInMainHand
-                    item?.modifyLore {
-                        clear()
-                    }
-                    sender.sendLang("Clear-Lore")
+                    if (item.isNotAir()) {
+                        item.modifyLore {
+                            clear()
+                        }
+                        sender.sendLang("Clear-Lore")
+                    } else sender.sendLang("Air-In-Hand")
                 }
             }
         }.dynamic("options") {
-            execute<ProxyCommandSender> { sender: ProxyCommandSender, context: CommandContext<ProxyCommandSender>, content: String ->
+            execute<ProxyCommandSender> { sender: ProxyCommandSender, _: CommandContext<ProxyCommandSender>, content: String ->
                 val args = CommandUtil.parseOptions(content.split(" "))
                 var silent = false
 
@@ -179,10 +212,12 @@ object LoreCommand {
 
                 sender.castSafely<Player>().let {
                     val item = it?.inventory?.itemInMainHand
-                    item?.modifyLore {
-                        clear()
-                    }
-                    if (!silent) sender.sendLang("Clear-Lore")
+                    if (item.isNotAir()) {
+                        item.modifyLore {
+                            clear()
+                        }
+                        if (!silent) sender.sendLang("Clear-Lore")
+                    } else sender.sendLang("Air-In-Hand")
                 }
             }
         }
@@ -192,10 +227,12 @@ object LoreCommand {
                 execute<ProxyCommandSender> { sender: ProxyCommandSender, context: CommandContext<ProxyCommandSender>, _: String ->
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            set(context.int("lineB") - 1, get(context.int("lineA") - 1))
-                        }
-                        sender.sendLang("Clone-Lore", context.int("lineA"), context.int("lineB"))
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                set(context.int("lineB") - 1, get(context.int("lineA") - 1))
+                            }
+                            sender.sendLang("Clone-Lore", context.int("lineA"), context.int("lineB"))
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }.dynamic("options") {
@@ -211,10 +248,12 @@ object LoreCommand {
 
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            set(context.int("lineB") - 1, get(context.int("lineA") - 1))
-                        }
-                        if (!silent) sender.sendLang("Clone-Lore", context.int("lineA"), context.int("lineB"))
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                set(context.int("lineB") - 1, get(context.int("lineA") - 1))
+                            }
+                            if (!silent) sender.sendLang("Clone-Lore", context.int("lineA"), context.int("lineB"))
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }
@@ -225,11 +264,13 @@ object LoreCommand {
                 execute<ProxyCommandSender> { sender: ProxyCommandSender, context: CommandContext<ProxyCommandSender>, _: String ->
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            val value = get(context.int("line") - 1)
-                            clipboard[it] = value
-                            sender.sendLang("Copy-Lore", context.int("line"), value)
-                        }
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                val value = get(context.int("line") - 1)
+                                clipboard[it] = value
+                                sender.sendLang("Copy-Lore", context.int("line"), value)
+                            }
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }.dynamic("options") {
@@ -245,11 +286,13 @@ object LoreCommand {
 
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            val value = get(context.int("line") - 1)
-                            clipboard[it] = value
-                            if (!silent) sender.sendLang("Copy-Lore", context.int("line"), value)
-                        }
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                val value = get(context.int("line") - 1)
+                                clipboard[it] = value
+                                if (!silent) sender.sendLang("Copy-Lore", context.int("line"), value)
+                            }
+                        } else sender.sendLang("Air-In-Hand")
                     }
 
                 }
@@ -261,14 +304,16 @@ object LoreCommand {
                 execute<ProxyCommandSender> { sender: ProxyCommandSender, context: CommandContext<ProxyCommandSender>, _: String ->
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            clipboard[it]?.colored()?.let { value ->
-                                if (clipboard.isNotEmpty()) {
-                                    set(context.int("line"), value)
-                                    sender.sendLang("Paste-Lore", context.int("line"), value)
-                                } else sender.sendLang("Clipboard-Empty")
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                clipboard[it]?.colored()?.let { value ->
+                                    if (clipboard.isNotEmpty()) {
+                                        set(context.int("line"), value)
+                                        sender.sendLang("Paste-Lore", context.int("line"), value)
+                                    } else sender.sendLang("Clipboard-Empty")
+                                }
                             }
-                        }
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }.dynamic("options") {
@@ -284,14 +329,16 @@ object LoreCommand {
 
                     sender.castSafely<Player>().let {
                         val item = it?.inventory?.itemInMainHand
-                        item?.modifyLore {
-                            clipboard[it]?.colored()?.let { value ->
-                                if (clipboard.isNotEmpty()) {
-                                    set(context.int("line"), value)
-                                    if (!silent) sender.sendLang("Paste-Lore", context.int("line"), value)
-                                } else if (!silent) sender.sendLang("Clipboard-Empty")
+                        if (item.isNotAir()) {
+                            item.modifyLore {
+                                clipboard[it]?.colored()?.let { value ->
+                                    if (clipboard.isNotEmpty()) {
+                                        set(context.int("line"), value)
+                                        if (!silent) sender.sendLang("Paste-Lore", context.int("line"), value)
+                                    } else if (!silent) sender.sendLang("Clipboard-Empty")
+                                }
                             }
-                        }
+                        } else sender.sendLang("Air-In-Hand")
                     }
                 }
             }
@@ -302,10 +349,12 @@ object LoreCommand {
             literal("check") {
                 execute<ProxyCommandSender> { sender: ProxyCommandSender, _: CommandContext<ProxyCommandSender>, _: String ->
                     sender.castSafely<Player>().let {
-                        sender.sendLang(
-                            "Clipboard-Check",
-                            clipboard[it] ?: sender.sendLang("Clipboard-Empty")
-                        )
+                        if (clipboard[it]?.isNotEmpty() == true) {
+                            sender.sendLang(
+                                "Clipboard-Check",
+                                clipboard[it]!!
+                            )
+                        } else sender.sendLang("Clipboard-Empty")
                     }
                 }
             }
@@ -346,7 +395,7 @@ object LoreCommand {
                     }
                 }
             }.dynamic("options") {
-                execute<ProxyCommandSender> { sender: ProxyCommandSender, context: CommandContext<ProxyCommandSender>, content: String ->
+                execute<ProxyCommandSender> { sender: ProxyCommandSender, _: CommandContext<ProxyCommandSender>, content: String ->
                     val args = CommandUtil.parseOptions(content.split(" "))
                     var silent = false
